app/requirements.txt
Django>=4.2.6,<5.0.0
flake8
pydocstyle
requests>=2.31.0,<3.0.0
psycopg2>=2.9.5,<3.0.0
django-constance[database]
django-json-widget>=1.1.1,<2.0.0
redis
django-redis
django-extensions
pytz
django-rest-framework==0.1.0
djangorestframework==3.15.1
faker
django-cors-headers>=4.3.0